--- a/package.json
+++ b/package.json
@@ -30,7 +30,7 @@
     "lerna": "^3.22.1",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.8.3",
-    "typescript": "~4.7.4"
+    "typescript": "~5.1.6"
   },
   "engines": {
     "node": "^16.13 || ^18.12"