--- a/packages/inter-protocol/src/proposals/startEconCommittee.js
+++ b/packages/inter-protocol/src/proposals/startEconCommittee.js
@@ -22,7 +22,7 @@ const sanitizePathSegment = name => {
 
 /**
  * @param {import('./econ-behaviors').EconomyBootstrapPowers} powers
- * @param {object} [config]
+ * @param {object} config
  * @param {object} [config.options]
  * @param {EconCommitteeOptions} [config.options.econCommitteeOptions]
  */