closes: #7534
supersedes: #6578

## Description

- chore: bump @typescript-eslint
- chore: bump typescript to 5.1
- chore(types): conform to 5.1 (annotations)
- chore(types): conform to 5.1 (assertions)

### Security Considerations

<!-- Does this change introduce new assumptions or dependencies that, if violated, could introduce security vulnerabilities? How does this PR change the boundaries between mutually-suspicious components? What new authorities are introduced by this change, perhaps by new API calls? 
-->

### Scaling Considerations

_slightly_ faster
`yarn lerna run --no-bail lint:types` times under:

this branch (5.1):
24.8s, 24.7s

master (4.7):
28.7s, 28.9s


### Documentation Considerations

This bumps to the RC, not release. For our purposes that's sufficient. We can bump to release when it's available. Taking this now let's us get the `satisfies` feature: `https://github.com/microsoft/TypeScript/pull/51753/`. 
### Testing Considerations

CI