@dckc 2023-05-22
> This bumps to the RC, not release. For our purposes that's sufficient.

I'm not clear on what our purposes are. Help?

> the satisfies feature: https://github.com/microsoft/TypeScript/pull/51753/.

I looked at that issue briefly; it's not clear to me how it helps us.

#7534 says
> We've been stuck on 4.7

I don't see a problem with 4.7.