@mhofman 2023-06-02
> I don't see a problem with 4.7.

In general we should keep up with newer revisions of TS. Besides bug fixes and improvements, it makes it more likely that the code editor will be using the same version in its language server (if using the default configuration to use the built-in version instead of repo specific version).